Move ObjectTests from test__field to test__bootstrapfield to match Object itself.
